Adjust the functions
