+++ src/lib/lumi.ts
@@ -1185,9 +1185,10 @@ function writeTime(buffer: Buffer, offset: number, time: number, isNextDay: bool
 /**
  * Formats a number of minutes into a user-readable 24-hour time notation in the form hh:mm.
  */
-function formatTime(timeMinutes: number): string {
+function formatTime(timeMinutes: number, padHour: boolean = false): string {
     const hours = Math.floor(timeMinutes / 60);
     const minutes = timeMinutes % 60;
+    if (padHour) return `${String(hours).padStart(2, "0")}:${String(minutes).padStart(2, "0")}`;
     return `${hours}:${String(minutes).padStart(2, "0")}`;
 }
 
@@ -1195,16 +1196,66 @@ function formatTime(timeMinutes: number): string {
  * Parses a 24-hour time notation string in the form hh:mm into a number of minutes.
  */
 function parseTime(timeString: string): number {
-    const parts = timeString.split(":");
+    const trimmedTimeString = timeString.trim();
+    if (!trimmedTimeString.match(/^\d{1,2}:\d{1,2}$/)) {
+        throw new Error(`Invalid time format: ${trimmedTimeString}. Expected HH:MM format (eg. "21:30")`);
+    }
+    const parts = trimmedTimeString.split(":");
 
     if (parts.length !== 2) {
-        throw new Error(`Cannot parse time string ${timeString}`);
+        throw new Error(`Cannot parse time string ${trimmedTimeString}`);
     }
 
     const hours = Number.parseInt(parts[0], 10);
     const minutes = Number.parseInt(parts[1], 10);
+    if (hours < 0 || hours > 23) {
+        throw new Error(`Invalid time format: ${trimmedTimeString}. Hours out of bounds - should be between 0 and 23.`);
+    }
+    if (minutes < 0 || minutes > 59) {
+        throw new Error(`Invalid time format: ${trimmedTimeString}. Minutes out of bounds - should be between 0 and 59.`);
+    }
 
     return hours * 60 + minutes;
+}
+
+// Time encoding & decoding (used eg in Aqara FP300) (0xMMHHmmhh)
+function encodeTimeFormat(startTime: string, endTime: string): number {
+    const start = parseTime(startTime);
+    const end = parseTime(endTime);
+    return Math.floor(start / 60) | ((start % 60) << 8) | (Math.floor(end / 60) << 16) | ((end % 60) << 24);
+}
+
+function decodeTimeFormat(value: number): { startTime: string, endTime: string } | null {
+    if (value < 0 || value > 0xffffffff) return null;
+
+    const startHour = value & 0xff;
+    const startMin = (value >>> 8) & 0xff;
+    const endHour = (value >>> 16) & 0xff;
+    const endMin = (value >>> 24) & 0xff;
+
+    if (startHour > 23 || startMin > 59 || endHour > 23 || endMin > 59) return null;
+
+    return {
+        startTime: formatTime(startHour * 60 + startMin),
+        endTime: formatTime(endHour * 60 + endMin)
+    };
+}
+
+// Encodes and decodes detection range composite value (used in FP300) from/to int
+function encodeDetectionRangeComposite(detection_range_value: number, rangeCount: number): { [key: string]: boolean } {
+    const composite_values: { [key: string]: boolean } = {};
+    for (let i = 0; i < rangeCount; ++i) {
+        composite_values[`detection_range_${i}`] = ((detection_range_value >>> i) & 1) == 1;
+    }
+    return composite_values;
+}
+
+function decodeDetectionRangeComposite(composite_values: { [key: string]: boolean }, rangeCount: number): number {
+    let intValue = 0;
+    for (let i = 0; i < rangeCount; ++i) {
+        if (composite_values[`detection_range_${i}`]) intValue |= 1 << i;
+    }
+    return intValue;
 }
 
 const stringifiedScheduleFragmentSeparator = "|";
@@ -1913,6 +1964,68 @@ export const lumiModernExtend = {
             reporting: false,
             ...args,
         }),
+    lumiLedDisabledNightTime: () => {
+        return {
+            isModernExtend: true,
+            exposes: [
+                e
+                    .text("schedule_start_time", ea.ALL)
+                    .withDescription(
+                        "LED disable schedule start time (HH:MM format)",
+                    ),
+                e
+                    .text("schedule_end_time", ea.ALL)
+                    .withDescription(
+                        "LED disable schedule end time (HH:MM format)",
+                    )
+            ],
+            fromZigbee: [
+                {
+                    cluster: "manuSpecificLumi",
+                    type: ["attributeReport", "readResponse"],
+                    convert: async (model, msg, publish, options, meta) => {
+                        if (msg.data["574"] !== undefined) {
+                            const rawValue = msg.data["574"];
+                            const decoded = decodeTimeFormat(rawValue);
+
+                            return {
+                                schedule_start_time: decoded ? decoded.startTime : '--:--',
+                                schedule_end_time: decoded ? decoded.endTime : '--:--',
+                                schedule_time_raw: rawValue
+                            }
+                        }
+                    },
+                },
+            ],
+            toZigbee: [
+                {
+                    key: ["schedule_start_time", "schedule_end_time"],
+                    convertSet: async (entity, key, value, meta) => {
+                        assertString(value);
+                        
+                        // Read current and replace the attribute being edited
+                        const newData: { [key: string]: any } = {
+                            schedule_start_time: meta.state?.schedule_start_time ?? "00:00",
+                            schedule_end_time: meta.state?.schedule_end_time ?? "00:00",
+                            schedule_time_raw: meta.state?.schedule_time_raw
+                        };
+                        newData[key] = value;
+
+                        // Encode and write
+                        const encodedValue = encodeTimeFormat(newData.schedule_start_time, newData.schedule_end_time);
+                        newData.schedule_time_raw = encodedValue;
+                        await entity.write("manuSpecificLumi", { 574: { value: encodedValue, type: 0x0023 } }, {manufacturerCode: manufacturerCode});
+
+                        return { state: newData };
+                    },
+                    convertGet: async (entity, key, meta) => {
+                        const endpoint = meta.device.getEndpoint(1);
+                        await endpoint.read("manuSpecificLumi", [0x023e], {manufacturerCode: manufacturerCode});
+                    },
+                },
+            ],
+        } satisfies ModernExtend;
+    },
     lumiButtonLock: (args?: Partial<modernExtend.BinaryArgs<"manuSpecificLumi">>) =>
         modernExtend.binary({
             name: "button_lock",
@@ -2283,6 +2396,126 @@ export const lumiModernExtend = {
             description: "Indicates whether the PIR sensor detects motion (in mmWave + PIR mode after mmWave presence detection PIR sensors gets turned off so this attribute might change to false although the presence is detected).",
         });
     },
+    fp300DetectionRange: (args?: {
+        range_offset: number,
+        ranges_count: number
+    }): ModernExtend => {
+        args = {
+            range_offset: 0.25,
+            ranges_count: 24,
+            ...args,
+        };
+
+        let detectionRangeComposite = e
+            .composite("detection_range_composite", "detection_range_composite", ea.ALL)
+            .withDescription("Specifies the detection range using set of boolean settings.");
+        for (let i = 0; i < args.ranges_count; ++i) {
+            detectionRangeComposite = detectionRangeComposite.withFeature(
+                e
+                    .binary(`detection_range_${i}`, ea.SET, true, false)
+                    .withDescription(`${(i * args.range_offset).toFixed(2)}m - ${(i * args.range_offset + 1).toFixed(2)}m`)
+            )
+        }
+
+        return {
+            isModernExtend: true,
+            exposes: [
+                e
+                    .numeric('detection_range', ea.ALL)
+                    .withValueMin(0)
+                    .withValueMax((1 << 24) - 1)
+                    .withValueStep(1)
+                    .withDescription("Specifies the range that is being detected. Requires mmWave radar mode. Press the on-device button to wake the device up and refresh its' settings."),
+                detectionRangeComposite
+            ],
+            fromZigbee: [
+                {
+                    cluster: "manuSpecificLumi",
+                    type: ["attributeReport", "readResponse"],
+                    convert: async (model, msg, publish, options, meta) => {
+                        if (msg.data["410"] && Buffer.isBuffer(msg.data["410"])) {
+                            const buffer = msg.data["410"];
+                            const detection_range_value = (buffer.length > 0) ? buffer.readUIntLE(2, 3) : 0xFFFFFF;
+
+                            return {
+                                detection_range_prefix: (buffer.length > 0) ? buffer.readUIntLE(0, 2) : 0x0300,
+                                detection_range: detection_range_value,
+                                detection_range_composite: encodeDetectionRangeComposite(detection_range_value, args.ranges_count)
+                            };
+                        }
+                    },
+                }
+            ],
+            toZigbee: [
+                {
+                    key: ["detection_range"],
+                    convertSet: async (entity, key, value, meta) => {
+                        assertNumber(value);
+
+                        const detection_range_prefix: number = (meta.state?.detection_range_prefix != null && typeof(meta.state?.detection_range_prefix) == 'number') ? meta.state?.detection_range_prefix : 0x0300;
+
+                        const buffer = Buffer.allocUnsafe(5);
+                        buffer.writeUIntLE(detection_range_prefix, 0, 2);
+                        buffer.writeUIntLE(value, 2, 3);
+
+                        await entity.write("manuSpecificLumi", {
+                            410: {value: buffer, type: 0x41}
+                        }, {manufacturerCode: manufacturerCode});
+                        return {
+                            state: {
+                                detection_range: value,
+                                detection_range_composite: encodeDetectionRangeComposite(value, args.ranges_count)
+                            }
+                        };
+                    },
+                    convertGet: async (entity, key, meta) => {
+                        const endpoint = meta.device.getEndpoint(1);
+                        await endpoint.read("manuSpecificLumi", [0x019a], {manufacturerCode: manufacturerCode});
+                    }
+                },
+                {
+                    key: ["detection_range_composite"],
+                    convertSet: async (entity, key, value, meta) => {
+                        assertObject(value);
+
+                        const detection_range_prefix: number = (meta.state?.detection_range_prefix != null && typeof(meta.state?.detection_range_prefix) == 'number') ? meta.state?.detection_range_prefix : 0x0300;
+                        const detection_range_value = decodeDetectionRangeComposite(value, args.ranges_count);
+                        
+                        const buffer = Buffer.allocUnsafe(5);
+                        buffer.writeUIntLE(detection_range_prefix, 0, 2);
+                        buffer.writeUIntLE(detection_range_value, 2, 3);
+
+                        await entity.write("manuSpecificLumi", { 410: {value: buffer, type: 0x41} }, {manufacturerCode: manufacturerCode});
+                        return {
+                            state: {
+                                detection_range: detection_range_value,
+                                detection_range_composite: value
+                            }
+                        };
+                    },
+                    convertGet: async (entity, key, meta) => {
+                        const endpoint = meta.device.getEndpoint(1);
+                        await endpoint.read("manuSpecificLumi", [0x019a], {manufacturerCode: manufacturerCode});
+                    }
+                },
+            ]
+        } satisfies ModernExtend;
+    },
+    fp300TrackDistance: () => {
+        return {
+            isModernExtend: true,
+            exposes: [e.enum("track_target_distance", ea.SET, ["Start Tracking Distance"]).withDescription("Initiate current target distance tracking.")],
+            toZigbee: [
+                {
+                    key: ["track_target_distance"],
+                    convertSet: async (entity, key, value, meta) => {
+                        // Uint8: 1 (0x08) attribute 0x0198 (attribute: 408)
+                        await entity.write("manuSpecificLumi", {408: {value: 1, type: 0x20}}, {manufacturerCode: manufacturerCode});
+                    },
+                },
+            ],
+        } satisfies ModernExtend;
+    },
     fp1ePresence: () => {
         const attribute = {ID: 0x0142, type: 0x20};
         return modernExtend.binary({
@@ -2357,9 +2590,6 @@ export const lumiModernExtend = {
                 },
             ],
         } satisfies ModernExtend;
-    },
-    fp300DetectionRange: () => {
-
     },
     lumiMultiClick: (args?: Partial<modernExtend.BinaryArgs<"manuSpecificLumi">>) =>
         modernExtend.binary({
